test(builds): exercise expiration deadlines through CLI
